refactor(codegen): extract a11y into plugin folder
